requirements.txt
google-api-python-client==2.99.0
google-auth==2.23.3
google-cloud-storage==2.12.0
google-cloud-secret-manager==2.18.0
pandas==2.1.1
matplotlib==3.8.0
fpdf==1.7.2
requests==2.31.0
flask==2.0.1
werkzeug==2.0.3
pytz==2023.3
mailjet-rest==1.3.4
plotly==5.18.0
weasyprint==52.5
jinja2==3.1.2
kaleido==0.2.1
beautifulsoup4==4.12.2